fix(UseSerializer): avoid Kotlin Reflection exception, streamline custom serialization usage
