+++ src/main/kotlin/dev/s7a/ktconfig/KtConfigSerializer.kt
@@ -1,10 +1,6 @@
 package dev.s7a.ktconfig
 
-import kotlin.reflect.KType
-
 interface KtConfigSerializer {
-    val type: KType
-
     fun deserialize(value: Any?): Any?
 
     fun serialize(value: Any?): Any?
